# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: PermsCraft REST API
  description: |
    Embedded REST API for the PermsCraft Minecraft permission plugin.

    ## Authentication
    Every request must include a valid API key in the `Authorization` header:
    ```
    Authorization: Bearer <api-key>
    ```
    Keys are managed in-game or from console:
    ```
    /pc apikey create <label> [scopes...]
    /pc apikey list
    /pc apikey revoke <label>
    ```

    ## Scopes
    | Scope    | Description |
    |----------|-------------|
    | `read`   | GET endpoints for groups, users, server info |
    | `write`  | POST/PUT/DELETE for groups and users |
    | `log`    | Access to audit log endpoints |
    | `sync`   | Trigger cache flush and sync operations |
    | `backup` | Export, import, and restore backups |
    | `admin`  | All scopes combined |

    ## Security
    > ⚠️ The API transmits keys over plain HTTP by default.
    > It is strongly recommended to use a TLS reverse proxy (nginx/Caddy)
    > and bind to `127.0.0.1` in `config.yml`.
    > See the [TLS setup guide](https://github.com/your-repo/wiki/REST-API-TLS).

    ## Rate Limiting
    Default: 120 requests per minute per IP. Configurable in `config.yml`.
  version: 1.0.0
  contact:
    name: PermsCraft
    url: https://github.com/your-repo
servers:
  - url: http://localhost:4567/api/v1
    description: Local server (default)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /sync/user/{uuid}:
    post:
      summary: Sync a specific user across servers
      tags: [Sync]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User sync triggered
  /sync/group/{name}:
    post:
      summary: Sync a specific group across servers
      tags: [Sync]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group sync triggered
...
  /cache/flush/user/{uuid}:
    post:
      summary: Flush cache for a specific user
      tags: [Sync]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User cache flushed
  /cache/flush/group/{name}:
    post:
      summary: Flush cache for a specific group
      tags: [Sync]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group cache flushed

  # ── Backup ──────────────────────────────────────────────────────────────────
...
  /server/players/{uuid}:
    get:
      summary: Get online player info
      tags: [Server]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Player info
        "404":
          description: Player not online
...
  /apikeys:
    get:
      summary: List API keys (labels and scopes only — no plaintext keys)
      tags: [API Keys]
      responses:
        "200":
          description: Key list

    post:
      summary: Create a new API key
      description: |
        The plaintext key is returned **once** in this response and never stored.
        Save it immediately.
      tags: [API Keys]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label:
                  type: string
                  example: discord-bot
                scopes:
                  type: array
                  items:
                    type: string
                  example: [read, log]
      responses:
        "201":
          description: Key created — plaintext shown once
          content:
            application/json:
              schema:
                type: object
                properties:
                  label:
                    type: string
                  key:
                    type: string
                    description: Plaintext key — store this, it won't be shown again
                  scopes:
                    type: array
                    items:
                      type: string
  /apikeys/{label}:
    delete:
      summary: Revoke an API key by label
      tags: [API Keys]
      parameters:
        - in: path
          name: label
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Key revoked
        "404":
          description: Key not found
...